adding handling for multiple args
